refactor: repository-wide code quality audit
Findings of a full sweep over the backend and the frontend, applied together:

- terminal.Start refuses a fork the provider cannot spell, and a fork with no
  conversation to branch. Both used to fall through resumeArgs and reopen the
  parent's own conversation under a second card. providers.SupportsFork was
  consulted by nothing in the product until now — the only gate was the
  frontend's copy of the list.
- store.reopen carries the pin across a park and resume; a session parked by
  `lich close` came back unpinned though nobody had unpinned it.
- The CORS preflight the RPC and the drop endpoint each spelled is one helper,
  rpc.Preflight.
- terminal.New's five transport closures are methods; relay.Observe is split
  into recordState and endedErrands; sessionCost reads the whole-cost providers
  through one selector and saves once; spawn.Close folds the keep and default
  paths.
- relay.go, terminal.go and cli.go were over the 800-line ceiling; their
  cohesive groups moved into observe/deliver/answer.go, start/io.go and
  cost/transport.go. No behaviour moved with them.
- Named the day window beside the weekly one; said why isOID reads 7 and 64;
  two comment forms staticcheck flagged.
- Tests for what had none: the fork guard, the carried pin, gitutil.NoPrompt,
  the terminal focus tick, panes-store's empty-project guards, and threadSlots'
  register/announce coalescing under jsdom.
