+++ internal/cli/cli.go
@@ -24,18 +24,13 @@
 package cli
 
 import (
-	"bytes"
-	"encoding/csv"
 	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
 	"io"
-	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"time"
 
@@ -44,7 +39,6 @@ import (
 	"github.com/omartelo/lich/internal/relay"
 	"github.com/omartelo/lich/internal/singleton"
 	"github.com/omartelo/lich/internal/spawn"
-	"github.com/omartelo/lich/internal/store"
 )
 
 // NotACommand is returned when the arguments name no lich subcommand at all,
@@ -664,130 +658,6 @@ func (c *client) worktrees(args []string) error {
 	return nil
 }
 
-func (c *client) cost(args []string) error {
-	flags := newFlagSet("cost")
-	project := flags.String("project", "", "only this project, by name")
-	provider := flags.String("provider", "", "only sessions running this provider")
-	since := flags.String("since", "", "only sessions active within this window, e.g. 7d or 12h")
-	asJSON := flags.Bool("json", false, "print the whole report as JSON")
-	asCSV := flags.Bool("csv", false, "print the per-project rows as CSV")
-	if err := c.parse(flags, args); err != nil {
-		return err
-	}
-	// One format or the other: asked for both, there is no answer to print that
-	// is not half of each.
-	if flags.NArg() != 0 || (*asJSON && *asCSV) {
-		return usageError("cost")
-	}
-	from, err := windowStart(*since)
-	if err != nil {
-		return err
-	}
-
-	var report store.CostReport
-	if err := c.call("store.CostTotals", []any{*project, *provider, from}, shortCall, &report); err != nil {
-		return err
-	}
-	report.Projects = asList(report.Projects)
-	switch {
-	case *asJSON:
-		return c.emit(report)
-	case *asCSV:
-		return c.emitCSV(report)
-	}
-	return c.printCost(report)
-}
-
-// windowStart turns a --since value into the unix second the report starts at,
-// or 0 for no window at all. Days are spelled out here because Go's duration
-// parser stops at hours and "what did this week cost" is the question the
-// command is asked.
-func windowStart(value string) (int64, error) {
-	if value == "" {
-		return 0, nil
-	}
-	span, err := parseSpan(value)
-	if err != nil || span <= 0 {
-		return 0, fmt.Errorf("--since takes a window like 7d, 24h or 90m, not %q", value)
-	}
-	return time.Now().Add(-span).Unix(), nil
-}
-
-func parseSpan(value string) (time.Duration, error) {
-	if days, ok := strings.CutSuffix(value, "d"); ok {
-		n, err := strconv.Atoi(days)
-		if err != nil {
-			return 0, err
-		}
-		return time.Duration(n) * 24 * time.Hour, nil
-	}
-	return time.ParseDuration(value)
-}
-
-// printCost writes the report as a table and then the line the total is never
-// printed without: what it leaves out. A sum over sessions lich could not price
-// is a lower bound, and one that says so is the difference between a number
-// worth acting on and a number that quietly reads as the whole bill.
-func (c *client) printCost(report store.CostReport) error {
-	if report.Sessions == 0 {
-		fmt.Fprintln(c.stdout, "No sessions matched.")
-		return nil
-	}
-	fmt.Fprintln(c.stdout, "project\tsessions\tunpriced\tcost")
-	for _, row := range report.Projects {
-		fmt.Fprintf(c.stdout, "%s\t%d\t%d\t%s\n", row.Project, row.Sessions, row.Unpriced, money(row.CostUSD))
-	}
-	fmt.Fprintf(c.stdout, "total\t%d\t%d\t%s\n", report.Sessions, report.Unpriced, money(report.CostUSD))
-	fmt.Fprintln(c.stdout, exclusion(report))
-	return nil
-}
-
-// exclusion words what the total does not cover. Off readout gets its own
-// sentence: it is the reason every session is unpriced, and without it a
-// machine that has simply never been asked to count reads as one that spent
-// nothing.
-func exclusion(report store.CostReport) string {
-	line := "Complete: every session in this total is priced."
-	if report.Unpriced > 0 {
-		line = fmt.Sprintf(
-			"Lower bound: %d unpriced of %d sessions — their spend is not in this total.",
-			report.Unpriced, report.Sessions,
-		)
-	}
-	if !report.Readout {
-		line += "\nThe cost readout is off, so nothing is being priced — turn on" +
-			` Settings › "Session readout in the footer" › "And cost".`
-	}
-	return line
-}
-
-// money is a figure as a command line shows it. Two places because that is what
-// a bill reads in; --json and --csv carry the exact number, which is what a
-// script adding these up needs.
-func money(usd float64) string {
-	return fmt.Sprintf("$%.2f", usd)
-}
-
-// emitCSV writes the per-project rows and no total: a total is the sum of the
-// columns, and the unpriced column travels beside the money, so what the number
-// leaves out survives the export rather than staying behind in lich.
-func (c *client) emitCSV(report store.CostReport) error {
-	rows := [][]string{{"project", "sessions", "unpriced", "cost_usd"}}
-	for _, row := range report.Projects {
-		rows = append(rows, []string{
-			row.Project,
-			strconv.Itoa(row.Sessions),
-			strconv.Itoa(row.Unpriced),
-			strconv.FormatFloat(row.CostUSD, 'f', 6, 64),
-		})
-	}
-	w := csv.NewWriter(c.stdout)
-	if err := w.WriteAll(rows); err != nil {
-		return fmt.Errorf("write the rows: %w", err)
-	}
-	return nil
-}
-
 // report prints a Send or Wait outcome. An answer goes to stdout on its own so
 // it can be read as the command's output; a wait that ran out says what to run
 // next, because an agent handed a bare ticket would have to guess.
@@ -838,195 +708,6 @@ func (c *client) emit(payload any) error {
 		return fmt.Errorf("encode the result: %w", err)
 	}
 	return nil
-}
-
-// call posts one RPC to the lich this session belongs to.
-func (c *client) call(method string, args []any, timeout time.Duration, out any) error {
-	port, token, err := c.coordinates()
-	if err != nil {
-		return err
-	}
-	body, err := json.Marshal(args)
-	if err != nil {
-		return fmt.Errorf("encode arguments: %w", err)
-	}
-
-	httpClient := &http.Client{Timeout: timeout}
-	status, payload, err := post(httpClient, endpoint(port, token, method), body)
-	if err != nil {
-		return err
-	}
-	// A refused token is not this caller's mistake — the coordinates it was given
-	// can have gone stale under it (see reissued). Retried once, on the same port,
-	// with the token the running instance recorded.
-	if status == http.StatusForbidden || status == http.StatusUnauthorized {
-		token, err = c.reissued(port, token)
-		if err != nil {
-			return err
-		}
-		if status, payload, err = post(httpClient, endpoint(port, token, method), body); err != nil {
-			return err
-		}
-	}
-	if status != http.StatusOK {
-		return fmt.Errorf("%s", failureOf(payload, status))
-	}
-	if out == nil {
-		return nil
-	}
-	if err := json.Unmarshal(payload, out); err != nil {
-		return fmt.Errorf("decode the reply: %w", err)
-	}
-	return nil
-}
-
-// post sends one RPC and reads the whole reply. The body is read here rather
-// than returned open because a refused call is sent a second time, and the
-// first response has to be finished with before the second one starts.
-func post(httpClient *http.Client, endpoint string, body []byte) (int, []byte, error) {
-	resp, err := httpClient.Post(endpoint, "application/json", bytes.NewReader(body))
-	if err != nil {
-		return 0, nil, fmt.Errorf("reach lich: %w", err)
-	}
-	defer resp.Body.Close()
-	payload, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return 0, nil, fmt.Errorf("read the reply: %w", err)
-	}
-	return resp.StatusCode, payload, nil
-}
-
-// endpoint builds the RPC URL for one call.
-func endpoint(port, token, method string) string {
-	return fmt.Sprintf(
-		"http://127.0.0.1:%s/rpc/%s?token=%s",
-		url.PathEscape(port), method, url.QueryEscape(token),
-	)
-}
-
-// reissued answers a listener that refused this call's token, with the token the
-// lich on that same port recorded — or with why there is none to try.
-//
-// The connect token is minted per launch and lives only in memory, so a lich
-// closed and opened again answers on the pinned port with a token no older
-// process can know. The coordinates in a PTY's environment are that older
-// process's, and a caller that outlives the session it was started in keeps
-// them: a background agent the harness parks in its own daemon, a nohup, a
-// detached pane. Every call it makes 403s from then on, forever, with nothing
-// on screen saying why — measured on a Claude Code background agent whose
-// `lich mcp` outlived a restart of the lich that spawned it.
-//
-// Only the same port is retried. The runtime file names one instance, and a
-// machine running a daily driver beside a `task dev` build has two: answering to
-// the other one would put this message in a window the caller cannot see, which
-// is the rule coordinates() follows too.
-func (c *client) reissued(port, refused string) (string, error) {
-	stale := fmt.Sprintf(
-		"lich refused this token on port %s: the coordinates in this environment "+
-			"were exported by an earlier lich, and the one running now cannot be "+
-			"reached with them", port,
-	)
-	if c.running == nil {
-		return "", fmt.Errorf("%s. Run this from a session of the running lich", stale)
-	}
-	info, err := c.running()
-	if err != nil {
-		return "", fmt.Errorf("%s, and the running instance could not be read: %w", stale, err)
-	}
-	if info == nil || info.Token == "" {
-		return "", fmt.Errorf("%s, and no running instance is recorded to ask instead", stale)
-	}
-	if strconv.Itoa(info.Port) != port {
-		return "", fmt.Errorf(
-			"%s. The lich recorded on this machine listens on port %d, and a message "+
-				"sent there would land in a window this caller cannot see — start it "+
-				"again from a session of that lich",
-			stale, info.Port,
-		)
-	}
-	if info.Token == refused {
-		return "", fmt.Errorf(
-			"lich refused this token on port %s, and it is the token the running "+
-				"instance recorded — so something other than lich is answering there",
-			port,
-		)
-	}
-	return info.Token, nil
-}
-
-// coordinates finds the lich to talk to: the one that spawned this PTY when
-// there is one, otherwise the instance running on this machine.
-//
-// The environment comes first on purpose. A session belongs to the lich that
-// spawned it, and on a machine running a daily driver beside a `task dev`
-// build, the runtime file names only one of them — answering to the wrong lich
-// would put a message in a window the caller cannot see.
-func (c *client) coordinates() (string, string, error) {
-	if port, token := c.env("LICH_PORT"), c.env("LICH_TOKEN"); port != "" && token != "" {
-		return port, token, nil
-	}
-	if c.running == nil {
-		return "", "", fmt.Errorf("no lich is running")
-	}
-	info, err := c.running()
-	if err != nil {
-		return "", "", err
-	}
-	if info == nil || info.Port == 0 || info.Token == "" {
-		return "", "", fmt.Errorf("no lich is running — open lich, or run this inside one of its sessions")
-	}
-	return strconv.Itoa(info.Port), info.Token, nil
-}
-
-// runningLich reads the loopback coordinates of the lich running on this
-// machine. It is the same runtime file install.sh reads to reach a running lich
-// from outside a session, and LICH_DEV selects the dev instance's own file just
-// as it does everywhere else.
-func runningLich() (*singleton.Info, error) {
-	dir, err := os.UserConfigDir()
-	if err != nil {
-		return nil, fmt.Errorf("resolve config directory: %w", err)
-	}
-	return singleton.Read(dir)
-}
-
-// sessionID is which card this command runs in, empty for a caller that has no
-// session — a script, a scheduled job, a plain shell. The relay words the
-// message it delivers differently for those, so an empty sender is a fact to
-// pass on rather than an error.
-func (c *client) sessionID() string {
-	return c.env("LICH_SESSION_ID")
-}
-
-// failureOf unwraps the RPC's {"error": "..."} envelope, falling back to the
-// HTTP status for a body that is not one.
-func failureOf(payload []byte, status int) string {
-	var envelope struct {
-		Error string `json:"error"`
-	}
-	if err := json.Unmarshal(payload, &envelope); err == nil && envelope.Error != "" {
-		return envelope.Error
-	}
-	return fmt.Sprintf("%d %s", status, http.StatusText(status))
-}
-
-// waitBudget is how long the client gives a call that blocks on another
-// session: the wait it asked for, plus room for the answer's trip back. A zero
-// timeout means the relay's own default.
-//
-// It is capped at the relay's own bound, and capped in seconds rather than in
-// Duration for the reason relay.waitFor gives: a number past about 9.2e9
-// overflows into a negative Duration, which http.Client reads as a deadline
-// already past — the POST fails on the spot and reports the send as unreachable
-// while the task it carried is being delivered.
-func waitBudget(seconds int) time.Duration {
-	if seconds <= 0 {
-		return relay.DefaultWait + callSlack
-	}
-	if seconds > relay.MaxWaitSeconds {
-		seconds = relay.MaxWaitSeconds
-	}
-	return time.Duration(seconds)*time.Second + callSlack
 }
 
 // newFlagSet builds a subcommand's flags. Its own reporting is silenced: a
