+++ internal/rpc/rpc.go
@@ -50,15 +50,26 @@ func (h *Handler) Deny(method string) {
 	h.denied[method] = true
 }
 
-// ServeHTTP handles one call. CORS is wide open on purpose: in dev the page's
-// origin is the Vite server, not this listener, and the random token —
-// enforced by the transport mount — is the actual auth, as with /ws.
-func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+// Preflight answers the CORS handshake every POST endpoint on the listener
+// shares, and reports whether it consumed the request. CORS is wide open on
+// purpose: in dev the page's origin is the Vite server, not this listener, and
+// the random token — enforced by the transport mount — is the actual auth, as
+// with /ws. Only the OPTIONS probe is answered here; the caller still gates
+// the method, because each endpoint spells its own refusal.
+func Preflight(w http.ResponseWriter, r *http.Request) bool {
 	w.Header().Set("Access-Control-Allow-Origin", "*")
 	w.Header().Set("Access-Control-Allow-Headers", "Content-Type")
 	w.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS")
-	if r.Method == http.MethodOptions {
-		w.WriteHeader(http.StatusNoContent)
+	if r.Method != http.MethodOptions {
+		return false
+	}
+	w.WriteHeader(http.StatusNoContent)
+	return true
+}
+
+// ServeHTTP handles one call.
+func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	if Preflight(w, r) {
 		return
 	}
 	if r.Method != http.MethodPost {
